// src/app/models/statement.ts
export class Statement {
  firstName: string;
  lastName: string;
  birthDate: Date;
  birthDay: number;
  birthMonth: number;
  birthYear: number;
  city: string;
  county: string;
  streetName: string;
  streetNo: number;
  building: string;
  buildingEntrance: string;
  floor: number;
  apartmentNo: number;
  destinations: string;
  date: Date;
  signatureSrc: string;
  reasonNumber: number;
}